fix(combat): resolve 6 compile errors against Minecraft 1.19.4 Fabric API
- Cast TrackedEntity.distance (double) to float at all three call sites
  (CombatEngine, SpacingController, AttackValidator) to fix lossy-conversion error
- Player.onGround() → Player.isOnGround() in AttackValidator (correct mojmap name)
- player.level() → ctx.world() in PearlController (player.level() doesn't exist
  in this mapping; ctx.world() returns the same Level and is already available)

+++ src/main/java/baritone/combat/CombatEngine.java
@@ -93,7 +93,7 @@ public final class CombatEngine {
         ThreatEntry target = targetSelector.select(awarenessCtx);
         if (target == null || !target.tracked.entity.isAlive()) return pause();
 
-        float distance = target.tracked.distance;
+        float distance = (float) target.tracked.distance;
 
         if (distance > ENGAGE_DISTANCE) {
             return new PathingCommand(
